refactor(provider): standardize helper function conventions across all resources
- Extract helper functions into companion <name>.go files for: annotation,
  asset, datahub_dataset, insight (new files); label, folder, user (extended)
- Remove unnecessary receivers from mapReportToModel and
  overlayReportComputedFields (pure data mapping, no r.client usage)
- Fix mapReportToModel parameter order from (state, resp) to (resp, state)
- Remove extra id parameter from report populateState (reads state.Id internally)
- Rename allocation overlayComputedFields to overlayAllocationComputedFields
- Remove dead Required-field assignments in overlayAllocationComputedFields
- Convert buildInsightRequest to plan.toInsightRequest() method
- Convert buildResourceResultsRequest to plan.toResourceResultsRequest() method
- Remove allowNotFound from report and allocation populateState
- Update skill docs: receiver rules, naming conventions, overlay pattern examples
