+++ internal/provider/folder.go
@@ -2,9 +2,44 @@
 package provider
 
 import (
+	"context"
+	"fmt"
+
 	"github.com/doitintl/terraform-provider-doit/internal/provider/models"
+	"github.com/hashicorp/terraform-plugin-framework/diag"
 	"github.com/hashicorp/terraform-plugin-framework/types"
 )
+
+// populateState fetches the folder from the API and populates the Terraform state.
+// On 404, state.Id is set to null to signal Terraform to remove the resource from state.
+func (r *folderResource) populateState(ctx context.Context, state *folderResourceModel) diag.Diagnostics {
+	folderResp, err := r.client.GetFolderWithResponse(ctx, state.Id.ValueString())
+	if err != nil {
+		return diag.Diagnostics{
+			diag.NewErrorDiagnostic("Error Reading Folder", "Could not read folder ID "+state.Id.ValueString()+": "+err.Error()),
+		}
+	}
+
+	if folderResp.StatusCode() == 404 {
+		state.Id = types.StringNull()
+		return nil
+	}
+
+	if folderResp.StatusCode() != 200 {
+		return diag.Diagnostics{
+			diag.NewErrorDiagnostic("Error Reading Folder", fmt.Sprintf("Unexpected status code %d for folder ID %s: %s", folderResp.StatusCode(), state.Id.ValueString(), string(folderResp.Body))),
+		}
+	}
+
+	if folderResp.JSON200 == nil {
+		return diag.Diagnostics{
+			diag.NewErrorDiagnostic("Error Reading Folder", "Received empty response body for folder ID "+state.Id.ValueString()),
+		}
+	}
+
+	mapFolderToModel(folderResp.JSON200, state)
+	return nil
+}
 
 // mapFolderToModel maps the API response to the Terraform model.
 // Used by Read and ImportState (full mapping from API response).
@@ -60,4 +95,42 @@ func overlayFolderComputedFields(apiResp *models.Folder, plan *folderResourceMod
 	}
 
 	// name: Required — never touch.
+}
+
+// toCreateRequest converts the TF model to a CreateFolderRequest.
+func (plan *folderResourceModel) toCreateRequest() models.CreateFolderRequest {
+	req := models.CreateFolderRequest{
+		Name: plan.Name.ValueString(),
+	}
+
+	if !plan.ParentFolderId.IsNull() && !plan.ParentFolderId.IsUnknown() {
+		req.ParentFolderId = new(plan.ParentFolderId.ValueString())
+	} else {
+		req.ParentFolderId = new("root")
+	}
+
+	if !plan.Description.IsNull() && !plan.Description.IsUnknown() {
+		req.Description = new(plan.Description.ValueString())
+	}
+
+	return req
+}
+
+// toUpdateRequest converts the TF model to an UpdateFolderRequest.
+func (plan *folderResourceModel) toUpdateRequest() models.UpdateFolderRequest {
+	req := models.UpdateFolderRequest{
+		Name: new(plan.Name.ValueString()),
+	}
+
+	if !plan.ParentFolderId.IsNull() && !plan.ParentFolderId.IsUnknown() {
+		req.ParentFolderId = new(plan.ParentFolderId.ValueString())
+	} else {
+		req.ParentFolderId = new("root")
+	}
+
+	if !plan.Description.IsNull() && !plan.Description.IsUnknown() {
+		req.Description = new(plan.Description.ValueString())
+	}
+
+	return req
 }
