+++ internal/provider/user.go
@@ -12,6 +12,25 @@ import (
 	openapi_types "github.com/oapi-codegen/runtime/types"
 )
 
+// populateState fetches the user from the API and populates the Terraform state.
+// On nil user (not found), state.Id is set to null to signal Terraform to remove the resource.
+func (r *userResource) populateState(ctx context.Context, state *userResourceModel) diag.Diagnostics {
+	email := state.Id.ValueString()
+
+	user, diags := r.lookupUser(ctx, email)
+	if diags.HasError() {
+		return diags
+	}
+
+	if user == nil {
+		state.Id = types.StringNull()
+		return nil
+	}
+
+	mapUserToModel(user, state)
+	return nil
+}
+
 // normalizePhone restores the "+" prefix that the API strips from phone
 // country codes. The spec accepts "^\+[0-9]+$" but stores without the "+".
 func normalizePhone(phone *string) types.String {
@@ -173,4 +192,58 @@ func (r *userResource) lookupUser(ctx context.Context, email string) (*models.Us
 
 	users := *listResp.JSON200.Users
 	return &users[0], diags
+}
+
+// toInviteRequest converts the TF model to an InviteUserRequest (Create path).
+func (plan *userResourceModel) toInviteRequest() models.InviteUserRequest {
+	req := models.InviteUserRequest{
+		Email: openapi_types.Email(plan.Email.ValueString()),
+	}
+
+	if !plan.FirstName.IsNull() && !plan.FirstName.IsUnknown() {
+		req.FirstName = plan.FirstName.ValueStringPointer()
+	}
+	if !plan.LastName.IsNull() && !plan.LastName.IsUnknown() {
+		req.LastName = plan.LastName.ValueStringPointer()
+	}
+	if !plan.JobTitle.IsNull() && !plan.JobTitle.IsUnknown() {
+		req.JobTitle = new(models.InviteUserRequestJobTitle(plan.JobTitle.ValueString()))
+	}
+	if !plan.RoleId.IsNull() && !plan.RoleId.IsUnknown() {
+		req.RoleId = plan.RoleId.ValueStringPointer()
+	}
+	if !plan.OrganizationId.IsNull() && !plan.OrganizationId.IsUnknown() {
+		req.OrganizationId = plan.OrganizationId.ValueStringPointer()
+	}
+
+	return req
+}
+
+// toUpdateRequest converts the TF model to an UpdateUserRequest (Update path).
+func (plan *userResourceModel) toUpdateRequest() models.UpdateUserRequest {
+	req := models.UpdateUserRequest{}
+
+	if !plan.FirstName.IsNull() && !plan.FirstName.IsUnknown() {
+		req.FirstName = plan.FirstName.ValueStringPointer()
+	}
+	if !plan.LastName.IsNull() && !plan.LastName.IsUnknown() {
+		req.LastName = plan.LastName.ValueStringPointer()
+	}
+	if !plan.JobTitle.IsNull() && !plan.JobTitle.IsUnknown() {
+		req.JobTitle = new(models.UpdateUserRequestJobTitle(plan.JobTitle.ValueString()))
+	}
+	if !plan.RoleId.IsNull() && !plan.RoleId.IsUnknown() {
+		req.RoleId = plan.RoleId.ValueStringPointer()
+	}
+	if !plan.Phone.IsNull() && !plan.Phone.IsUnknown() {
+		req.Phone = plan.Phone.ValueStringPointer()
+	}
+	if !plan.PhoneExtension.IsNull() && !plan.PhoneExtension.IsUnknown() {
+		req.PhoneExtension = plan.PhoneExtension.ValueStringPointer()
+	}
+	if !plan.Language.IsNull() && !plan.Language.IsUnknown() {
+		req.Language = new(models.UpdateUserRequestLanguage(plan.Language.ValueString()))
+	}
+
+	return req
 }
